feat: add new product from company page

+++ src/pages/ProductsList/utils/deleteProduct.ts
@@ -1,7 +1,8 @@
 import { db } from "../../../../firebase";
 import { doc, deleteDoc } from "firebase/firestore";
 
-const deleteProduct = async (id: string) => {
+const deleteProduct = async (id: string | undefined) => {
+  if (!id) return;
   await deleteDoc(doc(db, "products", id));
 };
